[Patches + PatchLOC + Hunks] Better new line handling.
Git can store content, that has other new lines like \r (CR). But git outputs lines, that are separated by a \r not as different lines, only as one line (e.g. "+ first line\rsecondline"). For the patch handling, it's importend, that lines are only splited at "\n", so that no MalformedLine (e.g. without a "+ " or "- " in the beginning) occurs.

Not sure, if this works on Windows, but with unix (linux, mac etc.) this should be fine.

Also added a catch for MalformedLine in PatchParser, to better know, if this occurs.