Bugfix for Command-Issue.
The last line of files that are not in UNIX format was not read by
the 'cat' method. One of the callback functions expected a '\n'
that never appeared.

This is a copy from repositoryhandler.
Original author: Luis Cañas Díaz <lcanas@libresoft.es>
Bug Spotted by Alexander Pepper.